### 2018-11-27 #269 @cboettig
Submission: emld
Transform Ecological Metadata Language (EML) files into JSON-LD
and back into EML.  Doing so creates a natural list-based representation
of EML in R, so that EML data can easily be manipulated using standard
R tools. This makes this package an effective backend for other tools
for working with EML.  Additioally, the JSON-LD representation enables
the use of developer-friendly JSON parsing and informatics-friendly
serializations such as RDF and SPARQL queries. This package is targeted
primarily at developers. By abstracting away the complexity of EML's
XML Schema, developers can build against intuitive list or JSON objects
and not have to worry about satisfying many of the additional constraints
of set by the schema (such as element ordering, which is handled automatically).

In practice, `emld` will basically be the backend machinery for a new release of the `EML`
package, which will retain largely the same user-facing API, with several improvements.
But both in terms of ease of reviewing and ease of package maintenance, I think it's important
to have these as separate packages.

```
Package: emld
Title: Ecological Metadata as Linked Data
Version: 0.0.2
Authors@R: person("Carl", "Boettiger",
                  email = "cboettig@gmail.com",
                  comment=c(ORCID = "http://orcid.org/0000-0002-1642-628X"))
Description: Transform Ecological Metadata Language (EML) files into JSON-LD
             and back into EML.  Doing so creates a natural list-based representation
             of EML in R, so that EML data can easily be manipulated using standard
             R tools. This makes this package an effective backend for other tools
             for working with EML.  Additioally, the JSON-LD representation enables
             the use of developer-friendly JSON parsing and informatics-friendly
             serializations such as RDF and SPARQL queries. This package is targeted
             primarily at developers. By abstracting away the complexity of EML's
             XML Schema, developers can build against intuitive list or JSON objects
             and not have to worry about satisfying many of the additional constraints
             of set by the schema (such as element ordering, which is handled automatically).
    covr,
    rdflib,
    jqr
Imports: xml2,
    jsonld,
    yaml

```

https://github.com/cboettig/emld

data publication, data retrieval, and reproducibility.

Researchers and ecological informatics teams (at NEON, LTER sites, NCEAS, etc) who are consuming or managing EML and data described by EML.

This overlaps with the `EML` package, and will soon be the backend to the newer version of said package.  This package works very differently than the current CRAN version of EML: instead of a massive S4 class system, this uses a single S3 class by exploiting the magic of JSON-LD.

  - Bryce Mecum, @amoeba
  - Peter Slaughter, @gothub
  - Kara Woo, @karawoo
  - Anna Krystalli, @annakrystalli
  - Kelly Hondula, @khondula
