+++ R/emld-package.R
@@ -1,4 +1,4 @@
 .onLoad  <- function(libname, pkgname) {
-  options("emld_db" = "eml-2.1.1")
+  options("emld_db" = "eml-2.2.0")
   invisible()
 }
