### 2018-11-29 48ff612
Add additional validation checks
