+++ R/eml_validate.R
@@ -30,6 +30,47 @@
 eml_validate <- function(eml,
                          encoding = "UTF-8",
                          schema = NULL) {
+
+  doc <- generalized_parser(eml, encoding = encoding)
+
+
+  # Use the EML namespace to find the EML version and the schema location
+  if (is.null(schema)) {
+    try(schema <- eml_locate_schema(doc))
+  }
+  schema_doc <- xml2::read_xml(schema)
+
+  ## Only do full validation if schemaLocation is eml.xml and not
+  ## merely a subclass
+  full <- grepl(".*eml.xsd$", paste0(strsplit(
+    xml2::xml_attr(xml2::xml_root(doc),
+    "schemaLocation"), "\\s+")[[1]], collapse = " "))
+
+  if(full){
+    result <- eml_additional_validation(doc, encoding = encoding)
+    if(!result) return(result)
+  }
+
+  result <- tryCatch(
+    xml2::xml_validate(doc, schema_doc),
+    error = function(e) {
+      warning("The document could not be validated.")
+      list(status = 1,
+           errors = c(NULL),
+           warnings = c(e))
+    }
+  )
+  result
+}
+
+
+# Return an xml_document representation of EML given any of the following
+# eml formats:
+#
+# 1. A path to file
+# 2. a list (emld object)
+# 3. an xml_document
+generalized_parser <- function(eml, encoding = "UTF-8"){
   if (is.character(eml)) {
     if (file.exists(eml)) {
       doc <- xml2::read_xml(eml, encoding = encoding)
@@ -40,30 +81,94 @@ eml_validate <- function(eml,
     ##  FIXME shouldn't have to write to tempfile,
     ## but  `doc <- as_xml(eml)` fails to drop xsi prefix on "schemaLocation"
     f <- tempfile()
-    x <- as_xml(eml, f)
     doc <- xml2::read_xml(f)
     unlink(f)
   } else {
     stop(paste("Did not recognize eml object with class", class(eml)))
   }
+  doc
+}
 
-  # Use the EML namespace to find the EML version and the schema location
-  if (is.null(schema)) {
-    try(schema <- eml_locate_schema(doc))
+
+
+# Not all EML validation rules can be inforced by the XSD definitions
+# here we implement the addtional checks described by:
+# https://github.com/NCEAS/eml/blob/BRANCH_EML_2_2/docs/eml-validation-refs.md
+eml_additional_validation <- function(eml,
+                                      encoding = "UTF-8") {
+
+
+  error_log <- character()
+  # Parse the XML document using an XML Schema-compliant parser
+  doc <- generalized_parser(eml, encoding = encoding)
+
+  # If the root element is not eml, then the document is invalid
+  if(xml2::xml_name(xml2::xml_root(doc), ns = xml_ns(doc)) == "eml")
+    error_log <- c(error_log, ("root element is not named 'eml'"))
+
+  # Elements which contain an annotation child element MUST contain an id attribute,
+  # unless the containing annotation element contains a references attribute
+  annot_without_ref <- xml2::xml_find_all(doc, "//annotation[not(references)]/..")
+  if(any(lapply(xml_attrs(annot_without_ref, "id"), length) == 0))
+    error_log <- c(error_log,
+                   paste("parent of any annotation must have id",
+                         "unless annotation contains a references attribute"))
+
+  ## If annotation has a references, parent cannot have ID
+  id_and_annotation <- xml2::xml_find_all(doc, "//*[@id]/annotation")
+  annotation_children <- xml2::xml_name(xml2::xml_children(id_and_annotation))
+  if("references" %in% annotation_children)
[…]
