### 2018-11-29 158c0a8
tidy up data-raw dir
experimental or stale files removed or moved to notebook
