+++ notebook/XSD-to-RDFS.R
@@ -1,3 +1,5 @@
+## EXPERIMENTAL ONLY
+
 library(xml2)
 library(purrr)
 
