### 2018-12-03 13de94d
more reasonable codecov targets
