+++ R/validate_units.R
@@ -1,0 +1,3 @@
+# Validate that every `standardUnit` is a standardUnit
+# Validate that any `customUnit` is defined in the customUnits
+#
