+++ codecov.yml
@@ -1,1 +1,16 @@
 comment: false
+
+coverage:
+  precision: 2
+  round: down
+  range: "65...100"
+
+  status:
+    project:
+      default:
+        threshold: 1%
+    patch:
+      default:
+        target: auto
+        threshold: 2%
+        only_pulls: true
