+++ tests/testthat/test-template.R
@@ -8,6 +8,8 @@ test_that("We can show slots for the creator object", {
 })
 
 test_that("template knows about internal classes too", {
+  skip("internal class template deprecated")
+  ## Only true in emld_db 2.1.1 right now... Probably deprecating this
   expect_output(print(template("ResponsibleParty")), "individualName: \\{\\}")
 })
 
