+++ tests/testthat/test-validator.R
@@ -4,7 +4,8 @@ testthat::test_that("we can use emld validator", {
   f <- system.file("extdata", "example.xml", package = "emld")
   testthat::expect_true(eml_validate(f))
   testthat::expect_true(eml_validate(xml2::read_xml(f)))
-  testthat::expect_true(eml_validate(as_emld(f)))
+  my_eml <- as_emld(f)
+  testthat::expect_true(eml_validate(my_eml))
 
   testthat::expect_error(eml_validate(f, schema = "notafile"))
 
