+++ codecov.yml
@@ -8,7 +8,7 @@ coverage:
   status:
     project:
       default:
-        threshold: 1%
+        threshold: 2%
     patch:
       default:
         target: auto
