### 2018-12-05 #269 @lmullen
Thanks for your submission, @cboettig. I have run the editor's checks below. Please address the requested changes below. I have begun looking for reviewers and I will update this thread with the reviewers and due date once all reviewers have agreed to take this on.

---

Changes for package authors:

- [ ] Please add an rOpenSci review badge to your repository. You can use this code:

- [ ] The package summary might be somewhat clearer if revised to make clearer what the words "natural," "intuitive," and so on mean. I took this to mean that XML is a pain to work with and R lists and JSON are easier to work with in R. While I agree with what it is not a universal opinion. I suggest taking another look at that language and bringing it together into a single sentence or clause that expresses the idea that it is more idiomatic to work with lists than XML in R.
- [ ] When you say that the package is a "backend for other tools for working with EML," does that mean other tools written in R?
- [ ] The copyright holder in the `LICENSE` file should be filled out or removed.
- [ ] Please correct the following failed test before the review begins.

```
test-template.R:11: failure: template knows about internal classes too
`print\(template\("ResponsibleParty"\)\)` does not match "individualName: \\{\\}".
Actual value: "\{\}"
```

- [ ] Please check the examples. When running `devtools::check(run_dont_test = TRUE)` I get the following error.

```
E  checking examples (619ms)
   Running examples in 'emld-Ex.R' failed
   The error most likely occurred in:

   > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
   > ### Name: eml_locate_schema
   > ### Title: eml_locate_schema
   > ### Aliases: eml_locate_schema
   >
   > ### ** Examples
   >
   > ## No test:
   > f <- system.file("examples", "example-eml-2.1.1.xml", package = "EML")
   > eml <- xml2::read_xml(f)
   Error: '' does not exist in current working directory ('/tmp/Rtmp2NUD8n/emld.Rcheck').
```

For reviewers:

- [ ] Please note whether you have any difficulties with the chain of dependencies (jsonld package -> V8 package -> V8 system libraries) that leads back to V8. This is at least a potential pain point.

- [ ] Check code that is not covered by the test suite. See the output of `goodpractice:gp()` and `covr::package_coverage()`.

```
  ✖ write unit tests for all functions, and all package code in general. 86%
    of code lines are covered by test cases.
```
