### 2018-12-05 #269 @lmullen
Two reviewers have agreed. Thanks to Kelly and Peter for being willing to take on this review. :tada:  Typically we ask reviewers to complete their review in 3 weeks, but in this case that includes several holidays and most people's winter break. So I have set the due date with that in mind. Kelly and Peter, in addition to the typical checks listed in the reviewers' guide, I've noted a couple of things to pay attention to in my editorial comments above.

Reviewers: @khondula and @gothub

Due date: 2019-01-04
