### 2018-12-05 d786856
Add additional validation checks (#26)
* Add additional validation checks

* tidy up data-raw dir

experimental or stale files removed or moved to notebook

* more reasonable codecov targets

* a bit more testing

* fix test

* validate units as well

* simpler example test

* Use RTOOLS in appveyor please!
