+++ R/eml_validate.R
@@ -30,29 +30,31 @@
 eml_validate <- function(eml,
                          encoding = "UTF-8",
                          schema = NULL) {
-  if (is.character(eml)) {
-    if (file.exists(eml)) {
-      doc <- xml2::read_xml(eml, encoding = encoding)
-    }
-  } else if (is(eml, "xml_document")) {
-    doc <- eml
-  } else if (is(eml, "list")){
-    ##  FIXME shouldn't have to write to tempfile,
-    ## but  `doc <- as_xml(eml)` fails to drop xsi prefix on "schemaLocation"
-    f <- tempfile()
-    x <- as_xml(eml, f)
-    doc <- xml2::read_xml(f)
-    unlink(f)
-  } else {
-    stop(paste("Did not recognize eml object with class", class(eml)))
-  }
+
+  doc <- generalized_parser(eml, encoding = encoding)
+
 
   # Use the EML namespace to find the EML version and the schema location
   if (is.null(schema)) {
     try(schema <- eml_locate_schema(doc))
   }
   schema_doc <- xml2::read_xml(schema)
-  result <- tryCatch(
+
+  ## Only do full validation if schemaLocation is eml.xml and not
+  ## merely a subclass
+  full <- grepl(".*eml.xsd$", paste0(strsplit(
+    xml2::xml_attr(xml2::xml_root(doc),
+    "schemaLocation"), "\\s+")[[1]], collapse = " "))
+
+  ## defaults
+  id_valid <- TRUE
+  unit_valid <- TRUE
+  if(full){
+    id_valid <- eml_additional_validation(doc, encoding = encoding)
+    unit_valid <- validate_units(doc, encoding = encoding)
+  }
+
+  schema_valid <- tryCatch(
     xml2::xml_validate(doc, schema_doc),
     error = function(e) {
       warning("The document could not be validated.")
@@ -61,9 +63,120 @@ eml_validate <- function(eml,
            warnings = c(e))
     }
   )
-
+  result <- schema_valid & id_valid & unit_valid
+  attr(result, "errors") <- c(attr(id_valid, "errors"),
+                              attr(unit_valid, "errors"),
+                              attr(schema_valid, "errors"))
   result
 }
+
+
+# Return an xml_document representation of EML given any of the following
+# eml formats:
+#
+# 1. A path to file
+# 2. a list (emld object)
+# 3. an xml_document
+generalized_parser <- function(eml, encoding = "UTF-8"){
+  if (is.character(eml)) {
+    if (file.exists(eml)) {
+      doc <- xml2::read_xml(eml, encoding = encoding)
+    }
+  } else if (is(eml, "xml_document")) {
+    doc <- eml
+  } else if (is.list(eml)){
+    ##  FIXME shouldn't have to write to tempfile,
+    ## but  `doc <- as_xml(eml)` fails to drop xsi prefix on "schemaLocation"
+    f <- tempfile()
+    as_xml(eml, f)
+    doc <- xml2::read_xml(f)
+    unlink(f)
+  } else {
+    stop(paste("Did not recognize eml object with class", class(eml)))
+  }
+  doc
+}
+
+
+
+# Not all EML validation rules can be inforced by the XSD definitions
+# here we implement the addtional checks described by:
+# https://github.com/NCEAS/eml/blob/BRANCH_EML_2_2/docs/eml-validation-refs.md
+eml_additional_validation <- function(eml,
+                                      encoding = "UTF-8") {
+
+
+  error_log <- character()
+  # Parse the XML document using an XML Schema-compliant parser
+  doc <- generalized_parser(eml, encoding = encoding)
+
+  # If the root element is not eml, then the document is invalid
+  if(xml2::xml_name(xml2::xml_root(doc), ns = xml_ns(doc)) == "eml")
[…]
