### 2018-12-06 02d03ac
update README, docs
