+++ README.Rmd
@@ -7,6 +7,7 @@ output: github_document
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/cboettig/emld?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/emld)
 [![Coverage Status](https://img.shields.io/codecov/c/github/cboettig/emld/master.svg)](https://codecov.io/github/cboettig/emld?branch=master)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/emld)](https://cran.r-project.org/package=emld)
+[![](https://badges.ropensci.org/269_status.svg)](https://github.com/ropensci/software-review/issues/269)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
