+++ DESCRIPTION
@@ -5,17 +5,19 @@ Authors@R: person("Carl", "Boettiger",
                   email = "cboettig@gmail.com", 
                   role = c("aut", "cre", "cph"),
                   comment=c(ORCID = "http://orcid.org/0000-0002-1642-628X"))
-Description: Transform Ecological Metadata Language (EML) files into JSON-LD
-             and back into EML.  Doing so creates a natural list-based representation
-             of EML in R, so that EML data can easily be manipulated using standard
-             R tools. This makes this package an effective backend for other tools
-             for working with EML.  Additionally, the JSON-LD representation enables
-             the use of developer-friendly JSON parsing and informatics-friendly
-             serializations such as RDF and SPARQL queries. This package is targeted
-             primarily at developers. By abstracting away the complexity of EML's
-             XML Schema, developers can build against intuitive list or JSON objects
-             and not have to worry about satisfying many of the additional constraints
-             of set by the schema (such as element ordering, which is handled automatically).
+Description: This is a utility for transforming Ecological Metadata Language
+        (EML) files into JSON-LD and back into EML.  Doing so creates a
+        list-based representation of EML in R, so that EML data can easily
+        be manipulated using standard R tools. This makes this package an
+        effective backend for other R-based tools  working with EML. By
+        abstracting away the complexity of EML's XML Schema, developers can
+        build around native R list objects and not have to worry about satisfying
+        many of the additional constraints of set by the schema (such as element
+        ordering, which is handled automatically). Additionally, the JSON-LD 
+        representation enables the use of developer-friendly JSON parsing and
+        serialization that may faciliate the use of EML in contexts outside of R,
+        as well as the informatics-friendly serializations such as RDF and
+        SPARQL queries.
 URL: https://github.com/cboettig/emld
 BugReports: https://github.com/cboettig/emld/issues
 Depends: R (>= 3.1.0)
