### 2018-12-06 #269 @cboettig
Thanks @lmullen !

- [x] badge added.
- [x] DESCRIPTION updated to avoid `natural` and `intuitive` and be a bit more precise:
```
        This is a utility for transforming Ecological Metadata Language
        (EML) files into JSON-LD and back into EML.  Doing so creates a
        list-based representation of EML in R, so that EML data can easily
        be manipulated using standard R tools. This makes this package an
        effective backend for other R-based tools  working with EML. By
        abstracting away the complexity of EML's XML Schema, developers can
        build around native R list objects and not have to worry about satisfying
        many of the additional constraints of set by the schema (such as element
        ordering, which is handled automatically). Additionally, the JSON-LD
        representation enables the use of developer-friendly JSON parsing and
        serialization that may faciliate the use of EML in contexts outside of R,
        as well as the informatics-friendly serializations such as RDF and
        SPARQL queries.
```
- [x] Also hopefully makes explicit what parts would be useful to other R developers (e.g. R lists) vs what would be useful outside of the R context (having a JSON serialization, which is convenient for, e.g. javascript / web developers)
- [x] License template is filled out now :-)
- [x] Fixed failing test and failing example.
