### 2019-01-07 #269 @gothub
##### For packages co-submitting to JOSS

- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
- [ ]The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:

Estimated hours spent reviewing: 5-6

---

## Reviewer Comments

The case made for the necessity of the `emld` package in converting EML to JSON-LD is clear and compelling. Providing the ability to inspect EML derived data as S3 lists, JSON and via SPARQL queries will facilitate creation of new tools and workflows to automate access and interpretation.

In the `Motivations` section of `README.md` the purpose and method of "extending EML with other semantic vocabularies"  isn't clear (to me). Consider adding a section such as "Extending EML" with an example, if this is within the scope of the package.

The help text for the `as_xml` `Description` contains only the text `as_xml` and does not describe what the function does. Also, the return type is not specified. The same is true for both `as_emld` and `as_json`, `template`.

There is no documentation index is available for the package.

There is no documentation for `?emld'` or `?emld-package`.

This may be outside the scope of the package, but no functions, strategies or examples are presented in the documentation for simple editing of EML data, i.e. (reading, simple edit, write out). This would be very useful, but if this functionality presents to much overlap with the `EML` package, pleas disregard this comment.

The `as_json`, `as_emld`, `as_xml` functions have a clear purpose and work as expected.

In addition to testing using the provided code samples, the following checks for a complex EML document were performed for a couple of EML documents such as https://goa.nceas.ucsb.edu/#view/urn:uuid:3249ada0-afe3-4dd6-875e-0f7928a4c171:

- verified that `as_json` produced valid JSON-LD
- verified that the following commands produced the original EML file (round trip):

```
x <- as_emld('metadata.xml') # using
as_xml(x, "new.xml")
```
#### R Source

The following potentially unresolved items exist, as show from a quick scan of the source code:

```
$ grep FIXME *
as_emld.R:    ## FIXME technically this assumes only our context
as_xml.R:## FIXME drop NAs too?
eml_validate.R:    ##  FIXME shouldn't have to write to tempfile,
eml_validate.R:  # FIXME technically must be unique inside parent system only...
emld-methods.R:## FIXME: Print method should drop context
```
#### Misc
devtools::check() reports non-standard file `LICENSE.md`
