### 2019-01-11 #269 @lmullen
Thank you, @khondula and @gothub for your thorough reviews.

@cboettig: Please let us know if you have any questions about the reviews. Will you be able to complete the necessary changes in two weeks time, i.e., by January 25?
