+++ NAMESPACE
@@ -13,7 +13,6 @@ S3method(print,emld)
 export(as_emld)
 export(as_json)
 export(as_xml)
-export(eml_locate_schema)
 export(eml_validate)
 export(template)
 importFrom(jsonld,jsonld_compact)
