+++ DESCRIPTION
@@ -15,7 +15,7 @@ Description: This is a utility for transforming Ecological Metadata Language
         many of the additional constraints of set by the schema (such as element
         ordering, which is handled automatically). Additionally, the JSON-LD 
         representation enables the use of developer-friendly JSON parsing and
-        serialization that may faciliate the use of EML in contexts outside of R,
+        serialization that may facilitate the use of EML in contexts outside of R,
         as well as the informatics-friendly serializations such as RDF and
         SPARQL queries.
 URL: https://github.com/cboettig/emld
@@ -25,6 +25,7 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
+Roxygen: list(markdown = TRUE)
 Suggests:
     spelling,
     testthat,
