### 2019-01-26 648e61b
address issues raised in rOpenSci review; closes #28, #29
