### 2019-01-26 13ee1d6
expand JOSS paper with examples from README
