### 2019-01-26 #269 @cboettig
@khondula @gothub @lmullen Thanks for these very helpful and detailed reviews!  I believe I have addressed all of these issues now.  I've made the edits in a separate branch and created a PR, https://github.com/cboettig/emld/pull/30 which I will leave un-merged for now, hopefully that provides an easy way to see the differences.  I've also summarized my replies to each of your point-by-point reviews (with a few clarifying questions and comments) in the two issues linked in that PR.  Anchors aweigh  :ship:  ⚓️!
