### 2019-01-26 #269 @lmullen
Thanks for making these changes, @cboettig.

@khondula and @gothub: Could you please review these changes and make any further comments within one week, by Feb 2? If the package is now satisfactory, you can check of any remaining relevant boxes in your main review.
