### 2019-02-01 #269 @gothub
@lmullen @cboettig I had a look at https://github.com/cboettig/emld/issues/29 and another look at the updated package and it looks great! Really useful contribution, thx!
