### 2019-02-04 #269 @khondula
@cboettig re: the `template` function - I think if I were more familiar with the EML package I would agree that it might fit there better, but it seems like a nice function to have since it might help a user get more familiar with eml if they are starting from scratch. The fxn documentation seems to still refer to it being used in the readme though ("see vignettes"), which might be confusing( unless I am missing a different vignette).

otherwise everything looks great to me!
