### 2019-02-05 ac5e3dd
update vignette, add NEWS
