+++ NEWS.md
@@ -1,0 +1,10 @@
+
+# emld 0.2.0
+
+* Implemented changes requested by rOpenSci review, as detailed in 
+  [#30](https://github.com/cboettig/emld/pull/30)
+
+# emld 0.1.1
+
+* Version submitted to rOpenSci review
+* Added a `NEWS.md` file to track changes to the package.
