### 2019-02-05 #269 @cboettig
@khondula @gothub Great!  Thanks to you both for the constructive reviews.  I've gone ahead and merged my own PR now.

@khondula
I rebuilt the [vignette](https://cboettig.github.io/emld/articles/tutorial.html), should have no mention of the `template()` function there now and I didn't see anything in searching for `template` in any of the main user facing functions.  It's still in the package as a bit of a vestigial organ for the time being, may either be flushed our or removed in favor of functions from `EML` package.

I think that's everything on my end for the moment; if you're happy with the little `paper.md` for the JOSS side?  Over to you, @lmullen
