### 2019-02-16 #269 @lmullen
Thank you @gothub and @khondula.

@cboettig We can consider this accepted now. Here are some post-review things for you to do.

- [ ] You are clear to submit the package to CRAN.
- [ ] Could you please transfer the repository to rOpenSci? Once you've done that I'll give you admin rights back.
- [ ] Please add an rOpenSci footer to the repository README. [![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
- [ ] Please add a CodeMeta file by running codemetar::write_codemeta()
- [ ] Please double check that CI and their respective badges work after the transfer.
