### 2019-02-16 #269 @lmullen
@cboettig Some other things pertaining to the JOSS submission.

- [ ] After the repository has been transferred, please generate a new release with a DOI. You may wish to wait until after CRAN acceptance.
- [ ] Please submit the paper to JOSS: http://joss.theoj.org/papers/new

Once you've submitted the paper, please let me know the URL to the submission and I will confirm that it has been peer reviewed by rOpenSci.
