### 2019-02-23 #269 @cboettig
Thanks @lmullen !  I've now completed the transfer to rOpenSci, and created the submission to JOSS at http://joss.theoj.org/papers/7e9064114c9a8d6756ef500cb5bc5d2b
