### 2019-02-26 #269 @lmullen
@cboettig Thanks, Carl. Looks like everything is wrapped up. Glad to see everything went through JOSS just fine.

@stefaniebutland This new package/paper might be a good candidate for a blog post at rOpenSci.
